requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.7.4.post0
altair==4.2.2
aniso8601==9.0.1
app==0.0.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.6.2
cachetools==5.3.0
Calculator==1.1.0
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==2.2.1
contourpy==1.0.7
convertdate==2.3.2
coverage==5.5
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
facexlib==0.3.0
fastjsonschema==2.15.3
filelock==3.12.0
filterpy==1.4.5
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
flatbuffers==23.3.3
fonttools==4.39.0
future==0.18.3
gast==0.4.0
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.2
imageio==2.26.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
ISR==2.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.2.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-tensorboard==0.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==15.0.6.1
llvmlite==0.39.1
lmdb==1.4.1
loadcaffe==0.1
lunardate==0.2.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
multidict==5.1.0
MyApplication==0.1.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
numba==0.56.4
numpy==1.22.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.4.20
pipreqs==0.4.13
platformdirs==3.5.0
pluggy==0.11.0
progressbar2==4.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.8.0
pyaml==21.10.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.14.0
pyluach==1.3.0
PyMeeus==0.5.11
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
-e git+https://github.com/xinntao/Real-ESRGAN.git@5ca1078535923d485892caee7d7804380bfc87fd#egg=realesrgan
requests==2.26.0
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scenedetect==0.6.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0.post1
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.22.0
super-resolution==0.2.4
tenacity==8.2.2
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.0
tensorflow-cpu==2.11.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-serving-api==2.11.0
tensorlayerx @ git+https://github.com/tensorlayer/tensorlayerx.git@be76ccf3a984214e1f7e6759b7c07520f84a1b1f
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
typing==3.7.4.3
typing-extensions==3.10.0.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.6
validators==0.20.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0
wget==3.2
widgetsnbextension==3.6.0
workalendar==16.0.0
wrapt==1.15.0
WTForms==2.3.3
yapf==0.33.0
yarg==0.1.9
yarl==1.6.3
zipp==3.15.0
re